Also add CVE ID to website
I have added the now published CVE (and GitHub's) ID of the vulnerability to the release at https://github.com/PrivateBin/PrivateBin/releases/tag/2.0.2 and this PR here is for the website, too.

OI guess this helps a lot for SEO etc. to find the corresponding information.

BTW @elrido the news does not seem to be published on https://privatebin.info/category/news.html yet.